requirements.txt
altgraph==0.15
appnope==0.1.0
awscli==1.14.34
beautifulsoup4==4.6.0
bleach==1.5.0
bokeh==0.12.11
boto3==1.5.24
botocore==1.8.38
bs4==0.0.1
cachetools==2.0.1
certifi==2018.1.18
chardet==3.0.4
clarifai==2.0.33
click==6.7
colorama==0.3.7
configparser==3.5.0
cycler==0.10.0
Cython==0.26.1
decorator==4.1.2
dicom==0.9.9.post1
Django==1.11.1
docutils==0.14
entrypoints==0.2.3
enum34==1.1.6
Flask==0.10.1
future==0.16.0
futures==3.1.1
get==1.0.1
google-api-python-client==1.6.5
google-auth==1.3.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
html5lib==0.9999999
httplib2==0.11.0
idna==2.6
image==1.5.16
imutils==0.4.3
inflection==0.3.1
ipdb==0.10.3
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.4
iso8601==0.1.12
itsdangerous==0.24
jedi==0.11.0
Jinja2==2.9.6
jmespath==0.9.3
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.4.0
linkedin-scraper==2.2.0
lxml==4.2.1
macholib==1.9
Markdown==2.6.9
MarkupSafe==0.23
matplotlib==2.0.2
mistune==0.8.1
more-itertools==3.2.0
nbconvert==5.3.1
nbformat==4.4.0
nose-for-sneeze==1.3.0
notebook==5.2.1
numpy==1.13.1
oauth2client==4.1.2
oauthlib==2.0.7
olefile==0.44
opencv-python==3.3.0.10
pandas==0.20.3
pandocfilters==1.4.2
parso==0.1.0
pefile==2017.11.5
pexpect==4.2.1
pickleshare==0.7.4
Pillow==2.9.0
plotly==2.2.2
post==1.0.0
prompt-toolkit==1.0.15
protobuf==3.4.0
ptyprocess==0.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyAudio==0.2.11
pycloudfn==0.1.209
Pygments==2.2.0
PyInstaller==3.3.1
pyparsing==2.2.0
pyserial==3.4
pyspin==1.1.1
python-dateutil==2.6.0
python-firebase==1.2
pytz==2017.2
PyYAML==3.12
pyzmq==16.0.3
qtconsole==4.3.1
Quandl==3.2.0
redis==2.10.6
request==1.0.0
requests==2.13.0
requests-oauthlib==0.8.0
rsa==3.4.2
s3transfer==0.1.12
scikit-learn==0.19.0
scipy==0.19.1
scrape-linkedin==0.1
selenium==3.11.0
simplegeneric==0.8.1
six==1.10.0
sklearn==0.0
tensorflow==1.3.0
tensorflow-gpu==1.1.0
tensorflow-tensorboard==0.1.6
terminado==0.6
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
virtualenv==15.1.0
wcwidth==0.1.7
Werkzeug==0.12
widgetsnbextension==3.0.7
